Fixes issue with json unmarshaling
Turns out the regexp filter doesn't work after and update. I dont
know regexp for shit so just removed the part that is not json.
This is not a permanent fix but guess that this will make fixing
the regexp very easy for a person who knows how regexp works